requirements.txt
streamlit>=1.32.0
supabase>=2.4.0
pandas>=2.0.0
python-dateutil>=2.8.2
pytz>=2023.3
resend>=0.7.0
python-dotenv>=1.0.0
